Fixed some issues, simplified and improved wishlist handling

+++ steam-wishlist-demos.py
@@ -1,5 +1,6 @@
 import time
 from enum import Enum
+import os.path
 import shutil
 import webbrowser
 
@@ -16,18 +17,17 @@ import requests
 
 class State(Enum):
   def __str__(self): return str(self.name)
-  Available = 0
+  Wished = 0
   Installed = 1
   Tried = 2
-  Missing = 3
-  Removed = 4
+  Failed = 3
 
 class Column(Enum):
   def __str__(self): return str(self.name)
-  AppId = 0
+  AppID = 0
   Name = 1
   State = 2
-  DemoId = 3
+  DemoID = 3
 
 NO_FILTER = '(any)'
 
@@ -36,7 +36,7 @@ data, data_sorted, data_sorted_filtered = {}, [], []
 
 layout = [
   [sg.Text('Steam profile ID:'), sg.Input('', key='SteamProfileId', size=(20,1)), sg.Text('Country code:'), sg.Input('', key='SteamCountryCode', size=(3,1))],
-  [sg.Text('Request:'), sg.Button('Wishlist'), sg.Button('Known demos', disabled=True), sg.Button('New demos', disabled=True), sg.Button('All demos', disabled=True),
+  [sg.Text('Request:'), sg.Button('Wishlist'), sg.Button('App-details', disabled=True),
     sg.ProgressBar(key='Progress', orientation='h', s=(10,20), expand_x=True, relief=sg.RELIEF_SUNKEN, max_value=100, visible=False),
     sg.Button('Stop', visible=False),],
   [sg.Text('Status:'), sg.Text('', key='ProgressText')],
@@ -46,19 +46,18 @@ layout = [
     sg.Button('Visit page', disabled=True),
     sg.Button('Install/Play demo', disabled=True)],
   [sg.Text('Modify the State of selected line(s):'),
-    sg.Button(State.Available, disabled=True),
+    sg.Button(State.Wished, disabled=True),
     sg.Button(State.Installed, disabled=True),
     sg.Button(State.Tried, disabled=True),
-    sg.Button(State.Missing, disabled=True),
-    sg.Button(State.Removed, disabled=True)],
+    sg.Button(State.Failed, disabled=True)],
   [sg.Text('Filters:'),
-    sg.Text('State ='), sg.Combo(key='FilterState', values=[NO_FILTER, State.Available, State.Installed, State.Tried, State.Missing, State.Removed], default_value=NO_FILTER, readonly=True, enable_events=True),
-    sg.Text('Has a DemoId ='), sg.Combo(key='FilterDemo', values=[NO_FILTER, 'Yes', 'No'], default_value=NO_FILTER, readonly=True, enable_events=True),
-    sg.Button('Clear all filters')],
+    sg.Text('State ='), sg.Combo(key='FilterState', values=[NO_FILTER, State.Wished, State.Installed, State.Tried, State.Failed], default_value=NO_FILTER, readonly=True, enable_events=True),
+    sg.Text('Has a DemoID ='), sg.Combo(key='FilterDemo', values=[NO_FILTER, 'Yes', 'No'], default_value=NO_FILTER, readonly=True, enable_events=True),
+    sg.Button('Reset')],
   [sg.Text('', key='TableTitle')],
   [sg.Table(key='Table',
             values=data_sorted_filtered,
-            headings=[Column.AppId, Column.Name, Column.State, Column.DemoId],
+            headings=[Column.AppID, Column.Name, Column.State, Column.DemoID],
             col_widths=[10, 40, 10, 10],
             num_rows=10, justification='left', def_col_width=10, auto_size_columns=False, expand_x=True, expand_y=True, change_submits=True)],
 ]
@@ -79,9 +78,9 @@ def check_filters(item):
       return False
   filterDemo = window['FilterDemo'].get()
   if filterDemo != NO_FILTER:
-    if filterDemo == 'Yes' and not item[Column.DemoId.value]:
+    if filterDemo == 'Yes' and not item[Column.DemoID.value]:
       return False
-    if filterDemo == 'No' and item[Column.DemoId.value]:
+    if filterDemo == 'No' and item[Column.DemoID.value]:
       return False
   return True
 
@@ -97,30 +96,27 @@ def update_table():
   num_filtered = len(data_sorted_filtered)
 
   if num_filtered != num_sorted:
-    num_demos_str = f"{num_filtered} (of {num_sorted})"
+    num_items_str = f"{num_filtered} (of {num_sorted})"
   else:
-    num_demos_str = f"{num_filtered}"
-
-  window['TableTitle'].update(f"Listing {num_demos_str} demos, from a total of {len(wishlist_appids)} wishlist items:")
+    num_items_str = f"{num_filtered}"
+  window['TableTitle'].update(f"Listing {num_items_str} wishlist items:")
 
   colored_rows = []
   for i, v in enumerate(data_sorted_filtered):
-      app_id = v[Column.AppId.value]
+      app_id = v[Column.AppID.value]
       state = v[Column.State.value]
-      demo_id = v[Column.DemoId.value]
-      if state == State.Removed:
-        colored_rows += [(i, 'white', 'black')]
-      elif not app_id in wishlist_appids:
+      demo_id = v[Column.DemoID.value]
+      if not app_id in wishlist_appids:
         colored_rows += [(i, 'white', 'red')]
-      elif state == State.Available and not demo_id:
+      elif state == State.Wished and not demo_id:
         colored_rows += [(i, 'black', 'orange')]
-      elif state == State.Available:
+      elif state == State.Wished:
         colored_rows += [(i, 'white', 'green')]
       elif state == State.Installed and not demo_id:
         colored_rows += [(i, 'black', 'yellow')]
       elif state == State.Tried and not demo_id:
         colored_rows += [(i, 'black', 'yellow')]
-      elif state == State.Missing:
+      elif state == State.Failed:
         colored_rows += [(i, 'white', 'gray')]
       else:
         colored_rows += [(i, 'black', 'white')]
@@ -134,30 +130,42 @@ backup_file = 'demos_installed-backup.txt'
 
 def load_data():
   global data
+
+  if not os.path.isfile(data_file):
+    return
+  
   print("Loading installed demos...")
   with open(data_file, 'r', encoding="utf-8") as file:
     lines = [line.rstrip() for line in file]
+
   if lines[0].startswith("steam_profile_id:"):
     _, steam_profile_id, steam_country_code = lines[0].split(':', 2)
     window['SteamProfileId'].update(steam_profile_id)
     window['SteamCountryCode'].update(steam_country_code)
+
   for line in lines[1:]:
     app_id_str, demo_id_str, prefix, name = line.split(':', 3)
     state = None
-    if prefix == 'A': state = State.Available
+    if prefix == 'W': state = State.Wished
     elif prefix == 'I': state = State.Installed
     elif prefix == 'T': state = State.Tried
-    elif prefix == 'M': state = State.Missing
+    elif prefix == 'F': state = State.Failed
     if state:
       app_id = int(app_id_str)
       demo_id = int(demo_id_str) if demo_id_str != 'None' else None
       data[app_id] = [app_id, name, state, demo_id]
+
   update_table()
+
+#--------------------------------------------------------------------------------
 
 def save_data():
   global data_sorted
-  print("Copying earlier installed demos to backup file...")
-  shutil.copyfile(data_file, backup_file)
+
+  if os.path.isfile(data_file):
+    print("Copying earlier installed demos to backup file...")
+    shutil.copyfile(data_file, backup_file)
+
   print("Saving installed demos...")
   with open(data_file, 'w', encoding="utf-8") as file:
     steam_profile_id = window['SteamProfileId'].get()
@@ -166,14 +174,14 @@ def save_data():
     for v in data_sorted:
       state = v[Column.State.value]
       prefix = None
-      if state == State.Available: prefix = 'A'
+      if state == State.Wished: prefix = 'W'
       elif state == State.Installed: prefix = 'I'
       elif state == State.Tried: prefix = 'T'
-      elif state == State.Missing: prefix = 'M'
+      elif state == State.Failed: prefix = 'F'
       if prefix:
         name = v[Column.Name.value]
-        app_id = v[Column.AppId.value]
-        demo_id = v[Column.DemoId.value]
+        app_id = v[Column.AppID.value]
+        demo_id = v[Column.DemoID.value]
         file.write(f"{app_id}:{demo_id}:{prefix}:{name}\n")
 
 #================================================================================
@@ -183,27 +191,56 @@ def get_wishlist(steam_profile_id):
 
   #response_items = steam.users.get_profile_wishlist(steam_profile_id)
   response = requests.get(f"https://api.steampowered.com/IWishlistService/GetWishlist/v1/", params={"steamid": steam_profile_id})
+
+  if not response.ok:
+    return False
+  
   response_items = response.json()['response']['items']
 
   wishlist_appids = [int(item['appid']) for item in response_items]
   print(f"Got {len(wishlist_appids)} apps for profile {steam_profile_id}")
+
+  for app_id in wishlist_appids:
+    row = data.get(app_id)
+    if not row:
+      data[app_id] = [app_id, "<Name is not fetched yet>", State.Wished, None]
+
   update_table()
+
+  missing_apps = [app_id for app_id in data.keys() if not app_id in wishlist_appids]
+  if len(missing_apps) > 0:
+    reply = sg.popup_yes_no(
+      f"Found {len(missing_apps)} AppIDs no longer on the wishlist.\nShould they be removed?\n(If not, they will just be marked in red.)",
+      title="Remove items?")
+    if reply == "Yes":
+      for app_id in missing_apps:
+        data.pop(app_id)
+      update_table()
+
+  return True
+
+#--------------------------------------------------------------------------------
 
 def request_wishlist():
   steam_profile_id = window['SteamProfileId'].get()
-  if steam_profile_id and steam_profile_id.isnumeric():
+  if not steam_profile_id:
+    window['ProgressText'].update(f"Enter a Steam profile ID first!", text_color='red')
+  elif not steam_profile_id.isnumeric():
+    window['ProgressText'].update(f"Steam profile ID must consist of only digits!", text_color='red')
+  else:
     print("Requesting wishlist...")
-    window['ProgressText'].update(f"Requesting wishlist...")
+    window['ProgressText'].update(f"Requesting wishlist...", text_color='white')
     window.refresh()
     try:
-      get_wishlist(steam_profile_id)
-      window['ProgressText'].update(f"Wishlist request completed!")
+      if get_wishlist(steam_profile_id):
+        window['ProgressText'].update(f"Wishlist request completed!", text_color='white')
+      else:
+        window['ProgressText'].update(f"Failed to request wishlist!", text_color='red')
     except Exception as e:
-      print(e)
-      window['ProgressText'].update(f"Exception ({type(e)}) occurred!")
-    window['Known demos'].update(disabled=(not wishlist_appids))
-    window['New demos'].update(disabled=(not wishlist_appids))
-    window['All demos'].update(disabled=(not wishlist_appids))
+      e_type = type(e).__name__
+      print(f"{e_type} '{e}'")
+      window['ProgressText'].update(f"{e_type} '{e}'", text_color='red')
+    window['App-details'].update(disabled=(not wishlist_appids))
 
 #================================================================================
 
@@ -230,25 +267,30 @@ def get_app_details(app_id, country_code):
     details = response.json()[str(app_id)]['data']
     name = details.get('name')
     demos = details.get('demos')
-    demo_appid_str = None
+
+    demo_id = None
     if demos:
-      demo_appid_str = ",".join([str(demo.get('appid')) for demo in demos if demo.get('appid')])
-    if demo_appid_str and ("," in demo_appid_str):
-      print(f"Warning: multiple demo-ids for app '{name}': {demo_appid_str}")
-    demo_id = int(demo_appid_str) if demo_appid_str else None
+      demo_ids = [demo.get('appid') for demo in demos if demo.get('appid')]
+      if len(demo_ids) > 0:
+        demo_id = demo_ids[0]
+      if len(demo_ids) > 1:
+        print(f"Warning: multiple demo-ids for app '{name}': {demo_ids}")
+
+    needs_update = False
     row = data.get(app_id)
     if row:
-      row[Column.Name.value] = name
-      if row[Column.DemoId.value] != demo_id:
-        print(f"Updating demo-id for app: '{name}'")
-        row[Column.DemoId.value] = demo_id
-        update_table()
-    elif demo_id:
-      data[app_id] = [app_id, name, State.Available, demo_id]
-      print(f"Found new app with demo: '{name}'")
+      if row[Column.Name.value] != name:
+        row[Column.Name.value] = name
+        print(f"Updated app Name: '{name}'")
+        needs_update = True
+      if row[Column.DemoID.value] != demo_id:
+        row[Column.DemoID.value] = demo_id
+        print(f"Updated demo-ID for app: '{name}'")
+        needs_update = True
+    if needs_update:
       update_table()
-    #print(f"{name} --> {demo_appid}")
     return True
+
   # Failure:
   print(f"App-details request failed, response status code: {response.status_code}")
   # 200 = successful
@@ -278,15 +320,10 @@ while True:
   if event == 'Wishlist':
     request_wishlist()
 
-  if event == 'Known demos' or event == 'New demos' or event == 'All demos':
+  if event == 'App-details':
+    fetch_appids = [id for id in wishlist_appids]
     window['Progress'].update(visible=True)
     window['Stop'].update(visible=True)
-    if event == 'Known demos':
-      fetch_appids = [id for id in wishlist_appids if id in data.keys()]
-    elif event == 'New demos':
-      fetch_appids = [id for id in wishlist_appids if not id in data.keys()]
-    elif event == 'All demos':
-      fetch_appids = [id for id in wishlist_appids]
     country_code = values['SteamCountryCode']
     print(f"Starting requests for app-details on '{event}' ({len(fetch_appids)} IDs, cc = '{country_code}')...")
     print(f"(Request interval: {request_interval_secs} secs)")
@@ -300,7 +337,7 @@ while True:
       #print(f"Requesting app-details for appid: {app_id}")
       remain_secs = avg_request_secs * (len(fetch_appids) - curr_app_idx - 1)
       m, s = int(remain_secs / 60), int(remain_secs) % 60
-      window['ProgressText'].update(f"Requesting demo for game {curr_app_idx+1}/{len(fetch_appids)}, time remaining: {m}m {s}s")
+      window['ProgressText'].update(f"Requesting app-details for item {curr_app_idx+1}/{len(fetch_appids)}, time remaining: {m}m {s}s", text_color='white')
       window.refresh()
       try:
         if get_app_details(app_id, country_code):
@@ -313,23 +350,24 @@ while True:
         else:
           # FIXME - allow failing on "too many requests", implement exponential back-off?
           print(f"Rate-limited, retrying after a delay of {retry_delay_secs} secs...")
-          window['ProgressText'].update(f"Rate-limited, waiting {retry_delay_secs} secs!")
+          window['ProgressText'].update(f"Rate-limited, waiting {retry_delay_secs} secs!", text_color='orange')
           window.timer_start(1000 * retry_delay_secs, repeating=False) # milliseconds
       except Exception as e:
-        print(e)
-        window['ProgressText'].update(f"Exception ({type(e)}) occurred, retrying in {retry_delay_secs} secs!")
+        e_type = type(e).__name__
+        print(f"{e_type} '{e}'")
+        window['ProgressText'].update(f"{e_type} '{e}' - retrying in {retry_delay_secs} secs!", text_color='red')
         window.timer_start(1000 * retry_delay_secs, repeating=False) # milliseconds
     else:
       print("App-details requests completed.")
       elapsed_secs = time.time() - start_time
       m, s = int(elapsed_secs / 60), int(elapsed_secs) % 60
       avg_secs = "{:.2f}".format(elapsed_secs / len(fetch_appids))
-      window['ProgressText'].update(f"Demo requests completed! (Time elapsed: {m}m {s}s, avg {avg_secs} secs)")
+      window['ProgressText'].update(f"App-detail requests completed! (Time elapsed: {m}m {s}s, avg {avg_secs} secs)", text_color='white')
       window['Progress'].update(visible=False)
       window['Stop'].update(visible=False)
       update_table()
       if len(fetch_appids) == 1:
-        selected_rows = [i for i, v in enumerate(data_sorted_filtered) if v[Column.AppId.value] == fetch_appids[0]]
+        selected_rows = [i for i, v in enumerate(data_sorted_filtered) if v[Column.AppID.value] == fetch_appids[0]]
         window['Table'].update(select_rows=selected_rows)
 
   if event == 'Stop':
@@ -337,7 +375,7 @@ while True:
       window.timer_stop_all()
       print("App-details requests stopped before completion!")
       # FIXME - display elapsed time?
-      window['ProgressText'].update(f"Demo requests were stopped!")
+      window['ProgressText'].update(f"Demo requests were stopped!", text_color='white')
       window['Progress'].update(visible=False)
       window['Stop'].update(visible=False)
       update_table()
@@ -345,7 +383,7 @@ while True:
   if event == 'FilterState' or event == 'FilterDemo':
     update_table()
 
-  if event == 'Clear all filters':
+  if event == 'Reset':
     window['FilterState'].update(NO_FILTER)
     window['FilterDemo'].update(NO_FILTER)
     update_table()
@@ -363,7 +401,7 @@ while True:
       idx = selected_rows[0]
       app_name = data_sorted_filtered[idx][Column.Name.value]
       window['Selection'].update(app_name)
-      demo_id = data_sorted_filtered[idx][Column.DemoId.value]
+      demo_id = data_sorted_filtered[idx][Column.DemoID.value]
       window['Refresh'].update(disabled=False)
       window['Visit page'].update(disabled=False)
       window['Install/Play demo'].update(disabled=(not demo_id))
@@ -377,10 +415,15 @@ while True:
 
   if selected_rows:
     needs_refresh = False
-    if event in [state.name for state in State]:
-      for idx in selected_rows:
-        data_sorted_filtered[idx][Column.State.value] = event
-      needs_refresh = True
+    for state in State:
+      if event == state.name:
+        for idx in selected_rows:
+          app_id = data_sorted_filtered[idx][Column.AppID.value]
+          row = data.get(app_id)
+          if row:
+            row[Column.State.value] = state
+        needs_refresh = True
+        break
     if needs_refresh:
       update_table()
       if values['FilterState'] == NO_FILTER and values['FilterDemo'] == NO_FILTER:
@@ -389,7 +432,7 @@ while True:
   if len(selected_rows) == 1:
     idx = selected_rows[0]
     if event == 'Refresh':
-      app_id = data_sorted_filtered[idx][Column.AppId.value]
+      app_id = data_sorted_filtered[idx][Column.AppID.value]
       fetch_appids = [app_id]
       country_code = values['SteamCountryCode']
       print(f"Starting request for app-details on appid {app_id} (cc = '{country_code}')...")
@@ -397,10 +440,10 @@ while True:
       start_time = time.time()
       window.timer_start(0, repeating=False) # milliseconds
     if event == 'Visit page':
-      app_id = data_sorted_filtered[idx][Column.AppId.value]
+      app_id = data_sorted_filtered[idx][Column.AppID.value]
       webbrowser.open(f"https://store.steampowered.com/app/{app_id}/")
     if event == 'Install/Play demo':
-      demo_id = data_sorted_filtered[idx][Column.DemoId.value]
+      demo_id = data_sorted_filtered[idx][Column.DemoID.value]
       if demo_id:
         webbrowser.open(f"steam://rungameid/{demo_id}")
